ops_scripts/requirements.txt
colorama
requests
tqdm
jinja2
Pillow
psutil
